:recycle: 所有版本都进行转换

+++ src/main/java/club/mcams/carpet/helpers/rule/maxBlockInteractionDistance/MaxInteractionDistanceMathHelper.java
@@ -26,10 +26,6 @@ import net.minecraft.util.math.MathHelper;
 
 public class MaxInteractionDistanceMathHelper {
     public static double getMaxSquaredReachDistance() {
-        //#if MC<11700
-        //$$ return MathHelper.square((float) AmsServerSettings.maxBlockInteractionDistance);
-        //#else
-        return MathHelper.square(AmsServerSettings.maxBlockInteractionDistance);
-        //#endif
+       return MathHelper.square((float) AmsServerSettings.maxBlockInteractionDistance);
     }
 }
